Move back part of profile subcommands to profile package
